Improve robustness, safety, and tests
Multiple fixes and improvements across the codebase: make several OllamaClient helpers package-visible for testing; pass AI request timeout from settings when pinging; harden DbSeedSettingsState by copying maps, returning unmodifiable views and enforcing defaults; make PendingUpdate and Row defensively copy/immut immutable collections; guard ProgressTracker against non-positive advances; preserve decimal bounds in SchemaIntrospector by switching min/max to Double; expand SQL reserved keyword set; switch TopologicalSorter to an iterative Tarjan implementation to avoid stack overflows and added a deep-chain test; make array formatting configurable and clean control characters in AbstractDialect, plus batch size guard; add DialectFactory caching; add logging and safer handling in ForeignKeyResolver (depth normalization, non-nullable-cycle warnings, changed behavior to defer/return null when appropriate) and add related tests; RowGenerator logs when it cannot generate full row counts and parses trimmed values; ValueGenerator and Column updated to use Double for min/max and convert to int as needed; RepetitionRule validates count >= 1; simplify RepetitionRulesPanel UI by removing prefix handling; DriverLoader centralizes driver dir handling, ensures directory creation, adds deregisterAll and classloader cleanup; NotificationHelper uses central notification constant; and many unit tests updated/added to cover these behaviors. These changes improve safety, immutability, configurability, and test coverage.

+++ src/main/java/com/luisppb16/dbseed/model/Column.java
@@ -59,8 +59,8 @@ public record Column(
     boolean uuid,
     int length,
     int scale,
-    Integer minValue,
-    Integer maxValue,
+    Double minValue,
+    Double maxValue,
     Set<String> allowedValues) {
 
   public Column {
